+++ src/frontend/web/src/components/Main/Modal.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import ChannelModal from './ChannelModal';
-import InviteModal from '../../../Common/InviteModal';
-import ModalBox from '../../../Common/ModalBox';
-import { ModalType } from '../../../../../types/modal.type';
+import ChannelModal from './Aside/Body/ChannelModal';
+import InviteModal from '../Common/InviteModal';
+import ModalBox from '../Common/ModalBox';
+import { ModalType } from '../../../types/modal.type';
 
 export default function Modal({
   type,
